+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added additional variants to `TextureUsage`.
 * Added field `sampler_index2` to `material.Texture`.
 * Added method `alpha_texture_channel_index` to `material.Material`.
+* Added type `material.MaterialFlags`.
 
 ### Changed
 * Changed numpy arrays to be row-major to properly match conventions.
@@ -25,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Renamed `shader_database.Dependency.constant` to `shader_database.Dependency.float`.
 * Renamed field `material.Material.m_unks2_2` to `m_unks2`.
 * Changed field `material.Material.alpha_test` to use type `material.Texture` to better match in game data.
+* Changed field `material.Material.flags` to use type `material.MaterialFlags` to better match in game data.
 
 ### Removed
 * Removed field `unk5` from `LodItem`.